+++ src/app/modules/select-records-page/select-records-page.component.ts
@@ -211,9 +211,10 @@ export class SelectRecordsPageComponent
   }
 
   /**
-   * Applies the variable table filter change.
+   * Loads variable records.
+   * @param pageNumber - page number to load
    */
-  filterVariables(): void {
+  loadVariables(pageNumber = 0): void {
     // TODO: Add paging which should be supported by CTSS.
     this.selectRecords.loadVariables(
       this.cart.getRecords('ResearchStudy'),
@@ -226,7 +227,8 @@ export class SelectRecordsPageComponent
             has_loinc: this.hasLoinc
           },
       this.variableTable?.filtersForm.value || {},
-      this.sort['Variable']
+      this.sort['Variable'],
+      pageNumber
     );
   }
 
@@ -249,7 +251,7 @@ export class SelectRecordsPageComponent
    */
   loadFirstPage(resourceType: string): void {
     if (resourceType === 'Variable') {
-      this.filterVariables();
+      this.loadVariables();
     } else {
       const sortParam = this.getSortParam(resourceType);
       // TODO: Currently, user can filter loaded ResearchStudy records on
@@ -258,6 +260,20 @@ export class SelectRecordsPageComponent
         resourceType,
         `$fhir/${resourceType}?_count=50${sortParam ? '&' + sortParam : ''}`
       );
+    }
+  }
+
+  /**
+   * Loads the next page of the specified resource type.
+   * @param resourceType - resource type.
+   */
+  loadNextPage(resourceType: string): void {
+    if (resourceType === 'Variable') {
+      this.loadVariables(
+        this.selectRecords.currentState[resourceType].currentPage + 1
+      );
+    } else {
+      this.selectRecords.loadNextPage(resourceType);
     }
   }
 }