+++ src/app/shared/select-records/select-records.service.ts
@@ -13,10 +13,15 @@ interface SelectRecordState {
   loading: boolean;
   // Array of loaded resources
   resources: Resource[];
-  // Next page URL
+  // Next page URL for regular FHIR resources
   nextBundleUrl?: string;
+  // Page number for CTSS variables
+  currentPage?: number;
   // Indicates whether we need to reload data
   reset?: boolean;
+  // The total number of records is used to determine whether the next page
+  // of CTSS variables exists
+  totalRecords?: number;
 }
 
 @Injectable({
@@ -119,6 +124,7 @@ export class SelectRecordsService {
    * @param params - http parameters
    * @param filters - filter values
    * @param sort - the current sort state
+   * @param pageNumber - page number to load
    */
   loadVariables(
     selectedResearchStudies: Resource[],
@@ -126,17 +132,25 @@ export class SelectRecordsService {
       [param: string]: any;
     },
     filters: any,
-    sort: Sort
+    sort: Sort,
+    pageNumber: number
   ): void {
     const resourceType = 'Variable';
-    // TODO: Add sorting when CTSS will support it
-    const currentState = {
-      loading: true,
-      resources: [],
-      progressValue: 0,
-      totalRecords: 0
-    };
-    this.currentState[resourceType] = currentState;
+    let currentState = this.currentState[resourceType];
+    if (currentState?.totalRecords <= pageNumber * 50) {
+      return;
+    }
+    if (pageNumber === 0) {
+      currentState = {
+        loading: true,
+        resources: [],
+        currentPage: pageNumber,
+        totalRecords: 0
+      };
+      this.currentState[resourceType] = currentState;
+    } else {
+      currentState.loading = true;
+    }
     const dataFields = {
       id: 'uid',
       display_name: 'display_name',
@@ -169,7 +183,8 @@ export class SelectRecordsService {
     this.resourceStream[resourceType] = this.http
       .get(url, {
         params: {
-          maxList: 50,
+          offset: pageNumber * 50,
+          count: 50,
           df: uniqDataFields.join(','),
           terms: '',
           q: query.join(' AND '),
@@ -189,6 +204,7 @@ export class SelectRecordsService {
         map((data: any) => {
           // TODO
           const total = data[0];
+          currentState.totalRecords = total;
           const list = data[3];
           if (total && list) {
             list.forEach((item) => {